incorporate Fred's new code into my aimedPatch function

+++ src/LCEdisperse.h
@@ -48,7 +48,7 @@ class LCE_Disperse_base: public virtual LifeCycleEvent
   /**The sex-specific dispersal matrices, [0] for males, [1] for females, might be used as connectivity matrix as well*/
   TMatrix* _DispMatrix[2]; 
   
-  vector< vector<unsigned int> > _reducedDispMat[2];
+  vector< vector<double> > _reducedDispMat[2];
 
   vector< vector<double> > _reducedDispMatProbs[2]; // Kim adding, this is the new matrix that can have rows of different sizes, i.e. is not a TMatrix
     // this should hold the dispersal probabilities of the patches in the kernel -- sum of the row must be 1. check with Fred b/c I want to make this just be 1 row, not same num rows as patches
